# lib/api-zod/src/generated/types/parseTasksRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ParseTasksRequest {
  /** Natural language description of tasks to do */
  input: string;
  /** Additional context about the project */
  context?: string;
}