# lib/api-zod/src/generated/types/parseTasksResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ParsedTask } from "./parsedTask";

export interface ParseTasksResponse {
  tasks: ParsedTask[];
  message: string;
}