# lib/api-zod/src/generated/types/parsedTask.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ParsedTaskPhase } from "./parsedTaskPhase";

export interface ParsedTask {
  title: string;
  phase: ParsedTaskPhase;
}